+++ package.json
@@ -6,6 +6,9 @@
     "start": "next start"
   },
   "dependencies": {
+    "@headlessui/react": "^1.4.2",
+    "@heroicons/react": "^1.0.5",
+    "heroicons-react": "^1.4.1",
     "next": "latest",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
