+++ pages/about-us.js
@@ -1,17 +1,12 @@
-import Head from 'next/head'
-import { Fragment } from 'react'
-import { Popover, Transition } from '@headlessui/react'
 import {
-  AnnotationIcon,
-  ChatAlt2Icon,
   InboxIcon,
-  MenuIcon,
-  QuestionMarkCircleIcon,
-  XIcon,
   SparklesIcon,
   CheckCircleIcon
 } from '@heroicons/react/outline'
-import { ChevronDownIcon } from '@heroicons/react/solid'
+
+import Footer from '../components/Footer'
+import WaitingList from '../components/WaitingList'
+import Header from '../components/Header'
 
 const stats = [
     { label: 'Founded', value: '2021' },
@@ -25,254 +20,16 @@ const includedFeatures = [
     'Top up when you want ',
     'Free consolidation service',
     'Trusted fund manager',
-  ]
-
-const solutions = [
-  {
-    name: 'Inbox',
-    description: 'Get a better understanding of where your traffic is coming from.',
-    href: '#',
-    icon: InboxIcon,
-  },
-  {
-    name: 'Messaging',
-    description: 'Speak directly to your customers in a more meaningful way.',
-    href: '#',
-    icon: AnnotationIcon,
-  },
-  { name: 'Live Chat', description: "Your customers' data will be safe and secure.", href: '#', icon: ChatAlt2Icon },
-  {
-    name: 'Knowledge Base',
-    description: "Connect with third-party tools that you're already using.",
-    href: '#',
-    icon: QuestionMarkCircleIcon,
-  },
 ]
 
 
-const navigation = [
-  { name: 'About Us', href: '/about-us' },
-  { name: 'Blog', href: '/blog' },
-  { name: 'FAQs', href: '/faqs' },
-]
 
-const navigationFooter = {
-    main: [
-      { name: 'About', href: '#' },
-      { name: 'Blog', href: '#' },
-      { name: 'Privacy Policy', href: '#' },
-    ],
-    social: [
-      {
-        name: 'Facebook',
-        href: '#',
-        icon: (props) => (
-          <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
-            <path
-              fillRule="evenodd"
-              d="M22 12c0-5.523-4.477-10-10-10S2 6.477 2 12c0 4.991 3.657 9.128 8.438 9.878v-6.987h-2.54V12h2.54V9.797c0-2.506 1.492-3.89 3.777-3.89 1.094 0 2.238.195 2.238.195v2.46h-1.26c-1.243 0-1.63.771-1.63 1.562V12h2.773l-.443 2.89h-2.33v6.988C18.343 21.128 22 16.991 22 12z"
-              clipRule="evenodd"
-            />
-          </svg>
-        ),
-      },
-      {
-        name: 'Instagram',
-        href: '#',
-        icon: (props) => (
-          <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
-            <path
-              fillRule="evenodd"
-              d="M12.315 2c2.43 0 2.784.013 3.808.06 1.064.049 1.791.218 2.427.465a4.902 4.902 0 011.772 1.153 4.902 4.902 0 011.153 1.772c.247.636.416 1.363.465 2.427.048 1.067.06 1.407.06 4.123v.08c0 2.643-.012 2.987-.06 4.043-.049 1.064-.218 1.791-.465 2.427a4.902 4.902 0 01-1.153 1.772 4.902 4.902 0 01-1.772 1.153c-.636.247-1.363.416-2.427.465-1.067.048-1.407.06-4.123.06h-.08c-2.643 0-2.987-.012-4.043-.06-1.064-.049-1.791-.218-2.427-.465a4.902 4.902 0 01-1.772-1.153 4.902 4.902 0 01-1.153-1.772c-.247-.636-.416-1.363-.465-2.427-.047-1.024-.06-1.379-.06-3.808v-.63c0-2.43.013-2.784.06-3.808.049-1.064.218-1.791.465-2.427a4.902 4.902 0 011.153-1.772A4.902 4.902 0 015.45 2.525c.636-.247 1.363-.416 2.427-.465C8.901 2.013 9.256 2 11.685 2h.63zm-.081 1.802h-.468c-2.456 0-2.784.011-3.807.058-.975.045-1.504.207-1.857.344-.467.182-.8.398-1.15.748-.35.35-.566.683-.748 1.15-.137.353-.3.882-.344 1.857-.047 1.023-.058 1.351-.058 3.807v.468c0 2.456.011 2.784.058 3.807.045.975.207 1.504.344 1.857.182.466.399.8.748 1.15.35.35.683.566 1.15.748.353.137.882.3 1.857.344 1.054.048 1.37.058 4.041.058h.08c2.597 0 2.917-.01 3.96-.058.976-.045 1.505-.207 1.858-.344.466-.182.8-.398 1.15-.748.35-.35.566-.683.748-1.15.137-.353.3-.882.344-1.857.048-1.055.058-1.37.058-4.041v-.08c0-2.597-.01-2.917-.058-3.96-.045-.976-.207-1.505-.344-1.858a3.097 3.097 0 00-.748-1.15 3.098 3.098 0 00-1.15-.748c-.353-.137-.882-.3-1.857-.344-1.023-.047-1.351-.058-3.807-.058zM12 6.865a5.135 5.135 0 110 10.27 5.135 5.135 0 010-10.27zm0 1.802a3.333 3.333 0 100 6.666 3.333 3.333 0 000-6.666zm5.338-3.205a1.2 1.2 0 110 2.4 1.2 1.2 0 010-2.4z"
-              clipRule="evenodd"
-            />
-          </svg>
-        ),
-      },
-    ],
-  }
-
-function classNames(...classes) {
-  return classes.filter(Boolean).join(' ')
-}
 
 export default function AboutUs() {
   return (
     <div className="main-container">
         <div className="bg-white">
-            <header>
-                <Popover className="relative bg-white">
-                <div className="flex justify-between items-center max-w-7xl mx-auto px-4 py-6 sm:px-6 md:justify-start md:space-x-10 lg:px-8">
-                    <div className="flex justify-start lg:w-0 lg:flex-1">
-                    <a href="#">
-                        <span className="sr-only">Workflow</span>
-                        <img
-                        className="h-8 w-auto sm:h-10"
-                        src="https://tailwindui.com/img/logos/workflow-mark-indigo-600.svg"
-                        alt=""
-                        />
-                    </a>
-                    </div>
-                    <div className="-mr-2 -my-2 md:hidden">
-                    <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
-                        <span className="sr-only">Open menu</span>
-                        <MenuIcon className="h-6 w-6" aria-hidden="true" />
-                    </Popover.Button>
-                    </div>
-                    <Popover.Group as="nav" className="hidden md:flex space-x-10">
-                    <Popover className="relative">
-                        {({ open }) => (
-                        <>
-                            <Popover.Button
-                            className={classNames(
-                                open ? 'text-gray-900' : 'text-gray-500',
-                                'group bg-white rounded-md inline-flex items-center text-base font-medium hover:text-gray-900 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500'
-                            )}
-                            >
-                            <span>Green Pensions</span>
-                            <ChevronDownIcon
-                                className={classNames(
-                                open ? 'text-gray-600' : 'text-gray-400',
-                                'ml-2 h-5 w-5 group-hover:text-gray-500'
-                                )}
-                                aria-hidden="true"
-                            />
-                            </Popover.Button>
-
-                            <Transition
-                            as={Fragment}
-                            enter="transition ease-out duration-200"
-                            enterFrom="opacity-0 translate-y-1"
-                            enterTo="opacity-100 translate-y-0"
-                            leave="transition ease-in duration-150"
-                            leaveFrom="opacity-100 translate-y-0"
-                            leaveTo="opacity-0 translate-y-1"
-                            >
-                            <Popover.Panel className="absolute z-10 -ml-4 mt-3 transform w-screen max-w-md lg:max-w-2xl lg:ml-0 lg:left-1/2 lg:-translate-x-1/2">
-                                <div className="rounded-lg shadow-lg ring-1 ring-black ring-opacity-5 overflow-hidden">
-                                <div className="relative grid gap-6 bg-white px-5 py-6 sm:gap-8 sm:p-8 lg:grid-cols-2">
-                                    {solutions.map((item) => (
-                                    <a
-                                        key={item.name}
-                                        href={item.href}
-                                        className="-m-3 p-3 flex items-start rounded-lg hover:bg-gray-50"
-                                    >
-                                        <div className="flex-shrink-0 flex items-center justify-center h-10 w-10 rounded-md bg-indigo-600 text-white sm:h-12 sm:w-12">
-                                        <item.icon className="h-6 w-6" aria-hidden="true" />
-                                        </div>
-                                        <div className="ml-4">
-                                        <p className="text-base font-medium text-gray-900">{item.name}</p>
-                                        <p className="mt-1 text-sm text-gray-500">{item.description}</p>
-                                        </div>
-                                    </a>
-                                    ))}
-                                </div>
-                                </div>
-                            </Popover.Panel>
-                            </Transition>
-                        </>
-                        )}
-                    </Popover>
-
-                    {navigation.map((item) => (
-                        <a key={item.name} href={item.href} className="text-base font-medium text-gray-500 hover:text-gray-900">
-                        {item.name}
-                        </a>
-                    ))}
-                    </Popover.Group>
-                    <div className="hidden md:flex items-center justify-end md:flex-1 lg:w-0">
-                    {/* <a href="#" className="whitespace-nowrap text-base font-medium text-gray-500 hover:text-gray-900">
-                        Sign in
-                    </a> */}
-                    <a
-                        href="#"
-                        className="ml-8 whitespace-nowrap inline-flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-indigo-600 hover:bg-indigo-700"
-                    >
-                        Sign up
-                    </a>
-                    </div>
-                </div>
-
-                <Transition
-                    as={Fragment}
-                    enter="duration-200 ease-out"
-                    enterFrom="opacity-0 scale-95"
-                    enterTo="opacity-100 scale-100"
-                    leave="duration-100 ease-in"
-                    leaveFrom="opacity-100 scale-100"
-                    leaveTo="opacity-0 scale-95"
-                >
-                    <Popover.Panel
-                    focus
-                    className="absolute z-30 top-0 inset-x-0 p-2 transition transform origin-top-right md:hidden"
-                    >
-                    <div className="rounded-lg shadow-lg ring-1 ring-black ring-opacity-5 bg-white divide-y-2 divide-gray-50">
-                        <div className="pt-5 pb-6 px-5">
-                        <div className="flex items-center justify-between">
-                            <div>
-                            <img
-                                className="h-8 w-auto"
-                                src="https://tailwindui.com/img/logos/workflow-mark-indigo-600.svg"
-                                alt="Workflow"
-                            />
-                            </div>
-                            <div className="-mr-2">
-                            <Popover.Button className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
-                                <span className="sr-only">Close menu</span>
-                                <XIcon className="h-6 w-6" aria-hidden="true" />
-                            </Popover.Button>
-                            </div>
-                        </div>
-                        <div className="mt-6">
-                            <nav className="grid grid-cols-1 gap-7">
-                            {solutions.map((item) => (
-                                <a
-                                key={item.name}
-                                href={item.href}
-                                className="-m-3 p-3 flex items-center rounded-lg hover:bg-gray-50"
-                                >
-                                <div className="flex-shrink-0 flex items-center justify-center h-10 w-10 rounded-md bg-indigo-600 text-white">
-                                    <item.icon className="h-6 w-6" aria-hidden="true" />
-                                </div>
-                                <div className="ml-4 text-base font-medium text-gray-900">{item.name}</div>
-                                </a>
-                            ))}
-                            </nav>
-                        </div>
-                        </div>
-                        <div className="py-6 px-5">
-                        <div className="grid grid-cols-2 gap-4">
-                            {navigation.map((item) => (
-                            <a
-                                key={item.name}
-                                href={item.href}
-                                className="text-base font-medium text-gray-900 hover:text-gray-700"
-                            >
-                                {item.name}
-                            </a>
-                            ))}
-                        </div>
-                        <div className="mt-6">
-                            <a
-                            href="#"
-                            className="w-full flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-indigo-600 hover:bg-indigo-700"
-                            >
-                            Sign up
-                            </a>
-                            <p className="mt-6 text-center text-base font-medium text-gray-500">
-                            Existing customer?
-                            <a href="#" className="text-gray-900">
-                                Sign in
-                            </a>
-                            </p>
-                        </div>
-                        </div>
-                    </div>
-                    </Popover.Panel>
-                </Transition>
-                </Popover>
-            </header>
-
+            <Header/>
             <main>
                 <div>
                 {/* Hero card */}
@@ -786,105 +543,8 @@ avoided</dt>
         </div>
       </div>
     </div>
-    <div className="bg-white py-16 sm:py-24">
-      <div className="relative sm:py-16">
-        <div aria-hidden="true" className="hidden sm:block">
-          <div className="absolute inset-y-0 left-0 w-1/2 bg-gray-50 rounded-r-3xl" />
-          <svg className="absolute top-8 left-1/2 -ml-3" width={404} height={392} fill="none" viewBox="0 0 404 392">
-            <defs>
-              <pattern
-                id="8228f071-bcee-4ec8-905a-2a059a2cc4fb"
-                x={0}
-                y={0}
-                width={20}
-                height={20}
-                patternUnits="userSpaceOnUse"
-              >
-                <rect x={0} y={0} width={4} height={4} className="text-gray-200" fill="currentColor" />
-              </pattern>
-            </defs>
-            <rect width={404} height={392} fill="url(#8228f071-bcee-4ec8-905a-2a059a2cc4fb)" />
-          </svg>
-        </div>
-        <div className="mx-auto max-w-md px-4 sm:max-w-3xl sm:px-6 lg:max-w-7xl lg:px-8">
-          <div className="relative rounded-2xl px-6 py-10 bg-indigo-600 overflow-hidden shadow-xl sm:px-12 sm:py-20">
-            <div aria-hidden="true" className="absolute inset-0 -mt-72 sm:-mt-32 md:mt-0">
-              <svg
-                className="absolute inset-0 h-full w-full"
-                preserveAspectRatio="xMidYMid slice"
-                xmlns="http://www.w3.org/2000/svg"
-                fill="none"
-                viewBox="0 0 1463 360"
-              >
-                <path
-                  className="text-indigo-500 text-opacity-40"
-                  fill="currentColor"
-                  d="M-82.673 72l1761.849 472.086-134.327 501.315-1761.85-472.086z"
-                />
-                <path
-                  className="text-indigo-700 text-opacity-40"
-                  fill="currentColor"
-                  d="M-217.088 544.086L1544.761 72l134.327 501.316-1761.849 472.086z"
-                />
-              </svg>
-            </div>
-            <div className="relative">
-              <div className="sm:text-center">
-                <h2 className="text-3xl font-extrabold text-white tracking-tight sm:text-4xl">
-                    Get notified when we&rsquo;re launching.
-                </h2>
-                <p className="mt-6 mx-auto max-w-2xl text-lg text-indigo-200">
-                    We are working on our app and before we publically launch we will reach out to our waiting list.
-                </p>
-              </div>
-              <form action="#" className="mt-12 sm:mx-auto sm:max-w-lg sm:flex">
-                <div className="min-w-0 flex-1">
-                  <label htmlFor="cta-email" className="sr-only">
-                    Email address
-                  </label>
-                  <input
-                    id="cta-email"
-                    type="email"
-                    className="block w-full border border-transparent rounded-md px-5 py-3 text-base text-gray-900 placeholder-gray-500 shadow-sm focus:outline-none focus:border-transparent focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600"
-                    placeholder="Enter your email"
-                  />
-                </div>
-                <div className="mt-4 sm:mt-0 sm:ml-3">
-                  <button
-                    type="submit"
-                    className="block w-full rounded-md border border-transparent px-5 py-3 bg-indigo-500 text-base font-medium text-white shadow hover:bg-indigo-400 focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-indigo-600 sm:px-10"
-                  >
-                    Yes, I want to invest in our planet
-                  </button>
-                </div>
-              </form>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <footer className="bg-white">
-      <div className="max-w-7xl mx-auto py-12 px-4 overflow-hidden sm:px-6 lg:px-8">
-        <nav className="-mx-5 -my-2 flex flex-wrap justify-center" aria-label="Footer">
-          {navigationFooter.main.map((item) => (
-            <div key={item.name} className="px-5 py-2">
-              <a href={item.href} className="text-base text-gray-500 hover:text-gray-900">
-                {item.name}
-              </a>
-            </div>
-          ))}
-        </nav>
-        <div className="mt-8 flex justify-center space-x-6">
-          {navigationFooter.social.map((item) => (
-            <a key={item.name} href={item.href} className="text-gray-400 hover:text-gray-500">
-              <span className="sr-only">{item.name}</span>
-              <item.icon className="h-6 w-6" aria-hidden="true" />
-            </a>
-          ))}
-        </div>
-        <p className="mt-8 text-center text-base text-gray-400">Biome Invest Ltd</p>
-      </div>
-    </footer>
+    <WaitingList/>
+    <Footer/>
     </div>
   )
 }
