+++ pages/faqs.js
@@ -1,10 +1,43 @@
 
 import { Disclosure } from '@headlessui/react'
 import { ChevronDownIcon } from '@heroicons/react/outline'
+import Footer from '../components/Footer'
+import WaitingList from '../components/WaitingList'
+import Header from '../components/Header'
 
 const faqs = [
   {
-    question: "What's the best thing about Switzerland?",
+    question: "What happens to my pension is Biome goes bust?",
+    answer:
+      "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
+  },
+  {
+    question: "Isn't it better to have multiple pension pots?",
+    answer:
+      "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
+  },
+  {
+    question: "What are the expected returns for green funds compared to normal funds?",
+    answer:
+      "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
+  },
+  {
+    question: "What are green pensions?",
+    answer:
+      "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
+  },
+  {
+    question: "How does Biome make money?",
+    answer:
+      "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
+  },
+  {
+    question: "What are the fees?",
+    answer:
+      "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
+  },
+  {
+    question: "Should I trust Biome with my pension?",
     answer:
       "I don't know, but the flag is a big plus. Lorem ipsum dolor sit amet consectetur adipisicing elit. Quas cupiditate laboriosam fugiat.",
   },
@@ -18,36 +51,41 @@ function classNames(...classes) {
 
 export default function Faqs() {
     return ( 
-        <div className="bg-gray-50">
-            <div className="max-w-7xl mx-auto py-12 px-4 sm:py-16 sm:px-6 lg:px-8">
-            <div className="max-w-3xl mx-auto divide-y-2 divide-gray-200">
-                <h2 className="text-center text-3xl font-extrabold text-gray-900 sm:text-4xl">Frequently asked questions</h2>
-                <dl className="mt-6 space-y-6 divide-y divide-gray-200">
-                {faqs.map((faq) => (
-                    <Disclosure as="div" key={faq.question} className="pt-6">
-                    {({ open }) => (
-                        <>
-                        <dt className="text-lg">
-                            <Disclosure.Button className="text-left w-full flex justify-between items-start text-gray-400">
-                            <span className="font-medium text-gray-900">{faq.question}</span>
-                            <span className="ml-6 h-7 flex items-center">
-                                <ChevronDownIcon
-                                className={classNames(open ? '-rotate-180' : 'rotate-0', 'h-6 w-6 transform')}
-                                aria-hidden="true"
-                                />
-                            </span>
-                            </Disclosure.Button>
-                        </dt>
-                        <Disclosure.Panel as="dd" className="mt-2 pr-12">
-                            <p className="text-base text-gray-500">{faq.answer}</p>
-                        </Disclosure.Panel>
-                        </>
-                    )}
-                    </Disclosure>
-                ))}
-                </dl>
+        <div className="main-container">
+        <Header/>
+            <div className="bg-gray-50">
+                <div className="max-w-7xl mx-auto py-12 px-4 sm:py-16 sm:px-6 lg:px-8">
+                <div className="max-w-3xl mx-auto divide-y-2 divide-gray-200">
+                    <h2 className="text-center text-3xl font-extrabold text-gray-900 sm:text-4xl">Frequently asked questions</h2>
+                    <dl className="mt-6 space-y-6 divide-y divide-gray-200">
+                    {faqs.map((faq) => (
+                        <Disclosure as="div" key={faq.question} className="pt-6">
+                        {({ open }) => (
+                            <>
+                            <dt className="text-lg">
+                                <Disclosure.Button className="text-left w-full flex justify-between items-start text-gray-400">
+                                <span className="font-medium text-gray-900">{faq.question}</span>
+                                <span className="ml-6 h-7 flex items-center">
+                                    <ChevronDownIcon
+                                    className={classNames(open ? '-rotate-180' : 'rotate-0', 'h-6 w-6 transform')}
+                                    aria-hidden="true"
+                                    />
+                                </span>
+                                </Disclosure.Button>
+                            </dt>
+                            <Disclosure.Panel as="dd" className="mt-2 pr-12">
+                                <p className="text-base text-gray-500">{faq.answer}</p>
+                            </Disclosure.Panel>
+                            </>
+                        )}
+                        </Disclosure>
+                    ))}
+                    </dl>
+                </div>
+                </div>
             </div>
-            </div>
+        <WaitingList/>
+        <Footer/>
         </div>
     )
 }
